requirements.txt
keras==2.3.1
tensorflow==2.1.0
h5py==2.10.0
requests==2.27.1
tqdm==4.64.0
numpy==1.21.6
Flask==2.1.2
Flask-Cors==3.0.10
protobuf==3.20.0
matplotlib==3.1.2
pillow==8.2.0
opencv-python-headless
sklearn